\name{kl_relabel-package}
\alias{kl_relabel-package}
\alias{kl_relabel}
\docType{package}
\title{\packageTitle{kl_relabel}}
\description{\packageDescription{kl_relabel}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{kl_relabel}
  \packageIndices{kl_relabel}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{kl_relabel}
  
  Maintainer: \packageMaintainer{kl_relabel}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
