% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lse}
\alias{lse}
\title{Log of the sum of exponential}
\usage{
lse(x)
}
\arguments{
\item{x}{a vector}
}
\value{
log of the sum of exponentiated elements in \code{x}
}
\description{
Log of the sum of exponential
}
