% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmsbm}
\alias{mmsbm}
\title{Posterior samples from mixed membership stochastic blockmodel}
\format{
A \code{stanfit} object:
\describe{
  \item{}{Model has two parameters: the membership vectors \code{pi} and the image matrix \code{theta}}
  \item{}{Two MCMC chains were run on simulated data using NUTS, each sampling 500 samples after 500 iterations of warmup}
  \item{}{The Stan-code can be accessed by typing \code{mmsbm@stanmodel}}
}
}
\usage{
mmsbm
}
\description{
A \code{stanfit} object obtained from fitting a mixed memebership stochastic
blockmodel to simulated data.
}
\keyword{datasets}
