% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.stanfit.R
\name{permute.stanfit}
\alias{permute.stanfit}
\title{Permute MCMC samples contained in a stanfit object using the KL-algorithm}
\usage{
permute.stanfit(fit, par.list, log.form = FALSE, maxit = 100,
  verbose = FALSE)
}
\arguments{
\item{fit}{a stanfit object}

\item{par.list}{a list, each of which elements is a character vector of
length 2. The first element of each entry of the list should be the
name of the parameter and the second the dimensions that should be
permuted (either "cols", "rows", or "both"). It is important that the
first entry should specify the memebership/classification matrix
based on which the KL-algorithm is run.}

\item{log.form}{a boolean that indicates whether the classification
probabilities are estimated on the log-scale}

\item{maxit}{maximum number of iterations to be used in the KL-algorithm}

\item{verbose}{boolean that indicates whether intermediate messages
should be printed out}
}
\value{
a stanfit object with posterior samples permuted
}
\description{
Permutes the samples stored within a stanfit object by relabeling them
using the KL-algorithm suggested by Stephens (2000)
}
