% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{relabel_kl_log}
\alias{relabel_kl_log}
\title{Relabel membership vector by minimizing KL-algorithm (log-scale)}
\usage{
relabel_kl_log(lphi, maxit = 100L, verbose = TRUE)
}
\arguments{
\item{lphi}{cube of length \code{S}, each element of which is a 
matrix of dimension N times K, where N is the number of individuals 
and K is the number of extreme types (or classes).}

\item{maxit}{the number of maximum iterations to run, defaults to 100.}

\item{verbose}{if TRUE, number of iterations and corresponding KL-distance 
values are printed.}
}
\value{
The function returns a Rcpp::List of three elements. 
    1) A cube, \code{permuted}, of the same dimensions as \code{phi} but 
    with the labels permuted; 2) \code{perms} is a \code{S} times \code{K} 
    matrix containing the permutations necessary to produce \code{permuted} 
    from \code{phi} (i.e., the mapping from \code{phi} to \code{permuted});
    and 3) the number of iterations run.
}
\description{
Relabels the membership vectors of a mixed membership model or 
mixed membership stochastic blockmodel, using the KL-algorithm
proposed by Stephens (2000).
}
\details{
OpenMP is enabled if available. Due to overhead, the inner-most
    loop is parallelized only if the number of latent classes/types
    is larger than 3.
}
