% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{relabel_true}
\alias{relabel_true}
\title{Relabel membership vector by minimizing KL-algorithm to true labels}
\usage{
relabel_true(phi, phi_true, verbose = FALSE)
}
\arguments{
\item{phi}{cube of length \code{S}, each element of which is a matrix of 
dimension \code{N} times \code{K}, where \code{N} is the number of 
individuals and K is the number of extreme types (or classes).}

\item{phi_true}{a \code{N} times \code{K} matrix containing the "true"
values of latent-class/mixed-membership probabilities}

\item{verbose}{if true, KL-divergence from true labels is calculated
and printed before and after relabeling}
}
\value{
Returns a Rcpp::List with two elements: 1) A arma::cube, 
  \code{permuted}, of the same dimensions as \code{phi} but 
  with the labels permuted; 2) \code{perms} is a \code{S} times \code{K} 
  matrix containing the permutations necessary to produce \code{permuted} 
  from \code{phi} (i.e., the mapping from \code{phi} to \code{permuted})
}
\description{
Relabels the membership vectors of a mixed membership model or 
mixed membership stochastic blockmodel, by minimizing the KL-divergence
to a priori known true labels.
}
\details{
OpenMP is enabled if available. Due to overhead, the inner-most
    loop is parallelized only if the number of latent classes/types
    is larger than 3.
}
