% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.output.R
\name{relabelMCMC}
\alias{relabelMCMC}
\title{Relabel MCMC output by KL-algorithm}
\usage{
relabelMCMC(x, maxit = 100, verbose = TRUE)
}
\arguments{
\item{x}{an \eqn{N \times K \times S} array of MCMC samples containing the
the assignment probabilities to latent classes/extreme types, where
 \eqn{N} is the number of units/individuals, \eqn{K} the number of
 latent classes, and \eqn{S} the number of samples}

\item{maxit}{number of maximum iterations to use in the algorithm}

\item{verbose}{if true, prints the intermediate results}
}
\value{
a list of four elements: \code{relabeled}, the relabeled
  array; \code{perms} the permutation pattern used for each sample;
  \code{iterations} the number of iterations that were needed to permute
  the array; and \code{status} with 0 = converged, 1 = not converged.
}
\description{
Relabel MCMC output from finite mixture models using Stephens KL algorithm
}
