% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.stanfit.R
\name{to_stan_array}
\alias{to_stan_array}
\title{Collapse long-form array of posterior sample into short-form}
\usage{
to_stan_array(x)
}
\arguments{
\item{x}{an array created by calling \code{extract_n_combine}}
}
\value{
an reshaped array which is callable by \code{stan::monitor}.
}
\description{
Collpases the long-form array created by \code{extract_n_combine} into
the array form that is created when calling \code{stan::extract} on a 
\code{stanfit} object with the options \code{permute = FALSE} and
\code{inc_warmup = FALSE}.
}
