% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.stanfit.R
\name{array_traceplot}
\alias{array_traceplot}
\title{Traceplot for Stan-form arrays}
\usage{
array_traceplot(x, par)
}
\arguments{
\item{x}{an array created by calling \code{extract_n_combine}}

\item{par}{name of the parameter to plot}
}
\value{
the function returns a \code{ggplot} object of the traceplot
}
\description{
Plots traceplot for Stan-form arrays
}
