% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.output.R
\name{permuteMCMC}
\alias{permuteMCMC}
\title{Relabel MCMC output based on Permutation Matrix}
\usage{
permuteMCMC(x, perms)
}
\arguments{
\item{x}{an array of MCMC samples, where the last index has to be equal
to the number of posterior samples, \code{S}, and the second-to-last
index equal to the number of latent classes/extreme types, \code{K}.}

\item{perms}{a \code{S} times \code{K} matrix of permutations needed to
transformed the raw MCMC output to the permuted output, e.g., the 
\code{perms} element from running \code{relabelMCMC}.}
}
\value{
\code{x} permuted according to \code{perms}
}
\description{
Relabel MCMC output from finite mixture models using the permutations 
found by some algorithm
}
\details{
Given a permutation that resolves the label-switching, this 
  function applies the mapping to another object. If the object to 
  permute is a three-dimensional array of dimension \code{N} times
  \code{K} times \code{S}, the function will treat \code{S} as the 
  number of posterior draws and will permute the columns of the sub-array
  which is of dimensions \code{N} times \code{K}. If \code{x} is a 
  matrix, the function assumes that \code{nrow(x) == S} and 
  \code{ncol(x) == K}.
}
