% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.output.R
\name{relabelMCMC}
\alias{relabelMCMC}
\title{Relabel MCMC output by KL-algorithm}
\usage{
relabelMCMC(x, maxit = 100, verbose = TRUE, log.p = TRUE)
}
\arguments{
\item{x}{an \code{S}\*\code{N}\*\code{K} array of MCMC samples containing
the the assignment probabilities to latent classes/extreme types, where
\code{N} is the number of units/individuals, \code{K} the number of
latent classes, and \code{S} the number of posterior samples}

\item{maxit}{number of maximum iterations to use in the algorithm}

\item{verbose}{if true, prints the intermediate results}

\item{log.p}{if TRUE, treats elements in x as log-probabilities}
}
\value{
Returns a list of four elements: \code{permuted}, the relabeled
  array; \code{perms} the permutation pattern used for each sample;
  \code{iterations} the number of iterations that were needed to permute
  the array; and \code{status} with 0 = converged, 1 = not converged.
}
\description{
Relabel MCMC output from finite mixture models using Stephens KL algorithm
}
