% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.output.R
\name{relabelTRUE}
\alias{relabelTRUE}
\title{Relabel MCMC output by minimizing KL-divergence to true assignment 
probabilities}
\usage{
relabelTRUE(x, x.true, verbose = TRUE, log.p = TRUE)
}
\arguments{
\item{x}{an \code{S}\*\code{N}\*\code{K} array of MCMC samples containing
the the assignment probabilities to latent classes/extreme types, where
\code{N} is the number of units/individuals, \code{K} the number of
latent classes, and \code{S} the number of posterior samples}

\item{x.true}{matrix of dimension \code{N}\*\code{K} which contains the 
true assignment/mixed-membership probabilities}

\item{verbose}{if true, prints KL-divergence to true probabilities before
and after relabeling}

\item{log.p}{if TRUE, treats elements in x as log-probabilities}
}
\value{
Returns a list of two elements: \code{permuted}, the relabeled
  array and, \code{perms}, the permutation pattern used for each sample
}
\description{
Relabels the membership vectors of a mixed membership model or 
mixed membership stochastic blockmodel, by minimizing the KL-divergence
to a priori known true labels.
}
